Added recording of shifts to a context. Background exec WIP
