+++ App.js
@@ -5,11 +5,14 @@ import { createStackNavigator } from '@react-navigation/stack';
 const Stack = createStackNavigator();
 import { UserProvider } from './src/context/UserProvider.js';
 import MainNavigator from './src/navigation/MainNavigator.js';
+import { ShiftTimerProvider } from './src/context/ShiftTimerContext.js';
 
 export default function App() {
   return (
     <UserProvider>
-      <MainNavigator />
+      <ShiftTimerProvider>
+        <MainNavigator />
+      </ShiftTimerProvider>
     </UserProvider>
   );
 }
