+++ package.json
@@ -16,9 +16,10 @@
     "async-storage": "^0.1.0",
     "asyncstorage": "^1.5.0",
     "expo": "~52.0.38",
+    "expo-notifications": "^0.29.14",
     "expo-status-bar": "~2.0.1",
-    "lodash.filter": "^4.6.0",
     "firebase": "^11.5.0",
+    "lodash.filter": "^4.6.0",
     "react": "18.3.1",
     "react-native": "0.76.7",
     "react-native-async-storage": "^0.0.1",
