+++ src/screens/AllShiftsScreen.js
@@ -47,6 +47,14 @@ export default function AllShiftsScreen() {
         }
         return "00:00:00"; // Fallback for invalid values
     };
+
+    const formatTime = (timeString) => {
+        if (!timeString) return "00:00";
+        const date = new Date(timeString);
+        const hours = ("0" + date.getHours()).slice(-2);
+        const minutes = ("0" + date.getMinutes()).slice(-2);
+        return `${hours}.${minutes}`;
+    }
     
     return (
         <View style={styles.container}>
@@ -61,8 +69,12 @@ export default function AllShiftsScreen() {
                     keyExtractor={(item, index) => index.toString()}
                     renderItem={({ item }) => (
                         <View style={styles.shiftItem}>
+                            {/* Shift Name */}
+                            <Text style={styles.shiftName}>{item.name|| "Undefined"}</Text>
+
+                            {/* Shift Details */}
                             <Text style={styles.shiftText}>
-                                {formatDate(item.startDate)} - {formatDate(item.endDate)}
+                                {formatTime(item.startTime)} - {formatTime(item.endTime)}
                             </Text>
                             <Text>Pvm: {formatDate(item.date)}</Text>
                             <Text>Kesto: {formatDuration(item.duration)}</Text>
